Show toolkit details when clicking any split branch, not just toolkit branches
