+++ ui/client/src/shared/tabKind.ts
@@ -1,6 +1,7 @@
 import { DocumentObj } from '../objects/DocumentObj';
 import { ProjectObj } from '../objects/ProjectObj';
-import { CENTRAL_REPO_FOLDER_ID, idFromDocId, idFromRepoId, idFromToolkitSplitId } from './idDocId';
+import { VALUE_GROUP_UNDEFINED } from '../utils/splitTree';
+import { CENTRAL_REPO_FOLDER_ID, idFromDocId, idFromRepoId, isSplitId, toolkitNameFromSplitId } from './idDocId';
 
 export enum TabKind {
   Notebook = 'notebook',
@@ -20,7 +21,7 @@ const isAgentDoc = (doc: DocumentObj) => {
 export const classifyTab = (showItemId: string, project: ProjectObj): TabKind | undefined => {
   if (showItemId === CENTRAL_REPO_FOLDER_ID) return TabKind.CentralRepository;
 
-  if (idFromToolkitSplitId(showItemId)) return TabKind.Toolkit;
+  if (isSplitId(showItemId)) return TabKind.Toolkit;
 
   if (idFromRepoId(showItemId)) return TabKind.Repository;
 
@@ -40,4 +41,21 @@ export const classifyTab = (showItemId: string, project: ProjectObj): TabKind | 
 export const tabKindClassName = (showItemId: string, project: ProjectObj): string | undefined => {
   const kind = classifyTab(showItemId, project);
   return kind ? `tab-kind-${kind}` : undefined;
+};
+
+// Returns a human-readable tab title for a given item ID.
+export const detailsTabName = (showItemId: string, project: ProjectObj): string => {
+  if (showItemId === CENTRAL_REPO_FOLDER_ID) return 'Repositories';
+  if (isSplitId(showItemId)) {
+    const toolkitName = toolkitNameFromSplitId(showItemId, project.documents);
+    if (toolkitName) return toolkitName === VALUE_GROUP_UNDEFINED ? 'No toolkit' : toolkitName;
+  }
+  const docid = idFromDocId(showItemId);
+  if (docid) {
+    const doc = project.allDocuments.find(d => d.docid === docid);
+    if (doc) return doc.isConfig ? project.name + ' config' : doc.name;
+  }
+  const repoPath = idFromRepoId(showItemId);
+  if (repoPath) return repoPath;
+  return project.name + ' config';
 };
